Development backend

See merge request jala-university1/cohort-3/oficial-pt-programa-o-4-cspr-244.ga.t2.25.m1/se-o-b/projeto-final/grupo-1/capstone-uno!14

+++ backend/src/services/gameService.js
@@ -1,6 +1,7 @@
 const getNextId = require("../utils/getNextId");
 const repository = require("../repositories/gameRepository");
 const jwt = require("jsonwebtoken");
+const AppError = require("../utils/appError");
 
 async function createGame(data) {
   const id = await getNextId("gameid");
@@ -17,14 +18,29 @@ async function createGame(data) {
 }
 
 async function getGame(id) {
-  return await repository.findGameById(id);
+  const game = await repository.findGameById(id);
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
+
+  return game;
 }
 
 async function updateGame(id, updates) {
+  const game = await repository.findGameById(id);
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
+
   return await repository.updateGameById(id, updates);
 }
 
 async function deleteGame(id) {
+  const game = await repository.findGameById(id);
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
+
   return await repository.deleteGameById(id);
 }
 
@@ -36,13 +52,13 @@ async function joinGame(gameId, accessToken) {
   const decoded = jwt.verify(accessToken, process.env.JWT_SECRET);
   const userId = decoded.id;
 
-  const game = await getGame(gameId);
+  const game = await repository.findGameById(gameId);
   if (!game) {
-    throw new Error("Game not found");
+    throw new AppError("Game not found", 404);
   }
 
   if (game.players.includes(userId)) {
-    throw new Error("User already in the game");
+    throw new AppError("User already in the game", 409);
   }
 
   game.players.push(userId);
@@ -56,11 +72,15 @@ async function startGame(gameId, accessToken) {
 
   const game = await repository.findGameById(gameId);
   if (!game) {
-    throw new Error("Game not found");
+    throw new AppError("Game not found", 404);
   }
 
   if (game.creator !== userId) {
-    throw new Error("Only the game creator can start the game");
+    throw new AppError("Only the game creator can start the game", 403);
+  }
+
+  if (game.players.length < 2) {
+    throw new AppError("Insufficient number of players", 400);
   }
 
   if (game.players.length < 2) {
@@ -72,7 +92,7 @@ async function startGame(gameId, accessToken) {
   );
 
   if (!allReady) {
-    throw new Error("Not all players are ready");
+    throw new AppError("Not all players are ready", 400);
   }
 
   const updatedGame = await repository.updateGameById(gameId, {
@@ -89,15 +109,15 @@ async function markAsReady(gameId, accessToken) {
 
   const game = await repository.findGameById(gameId);
   if (!game) {
-    throw new Error("Game not found");
+    throw new AppError("Game not found", 404);
   }
 
   if (!game.players.includes(userId)) {
-    throw new Error("User not in game");
+    throw new AppError("User not in game", 400);
   }
 
   if (game.readyPlayers.includes(userId)) {
-    throw new Error("User already has been marked as ready");
+    throw new AppError("User already has been marked as ready", 400);
   }
 
   game.readyPlayers.push(userId);
@@ -110,18 +130,20 @@ async function leaveGame(gameId, accessToken) {
   const userId = decoded.id;
 
   const game = await repository.findGameById(gameId);
-  if (!game) throw new Error("Game not found");
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
 
   if (game.status !== "active" && game.status !== "not_started") {
-    throw new Error("Game is not in progress");
+    throw new AppError("Game is not in progress", 400);
   }
 
   if (!game.players.includes(userId)) {
-    throw new Error("User not in the game");
+    throw new AppError("User not in the game", 400);
   }
 
   if (game.leftPlayers.includes(userId)) {
-    throw new Error("User already left the game");
+    throw new AppError("User already left the game", 400);
   }
 
   game.leftPlayers.push(userId);
@@ -142,10 +164,12 @@ async function endGame(gameId, accessToken) {
   const userId = decoded.id;
 
   const game = await repository.findGameById(gameId);
-  if (!game) throw new Error("Game not found");
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
 
   if (game.creator !== userId) {
-    throw new Error("Only the game creator can end the game");
+    throw new AppError("Only the game creator can end the game", 401);
   }
 
   if (game.status !== "active") {
@@ -159,7 +183,9 @@ async function endGame(gameId, accessToken) {
 
 async function getGameState(gameId) {
   const game = await repository.findGameById(gameId);
-  if (!game) throw new Error("Game not found");
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
 
   return {
     game_id: game.id,
@@ -169,7 +195,9 @@ async function getGameState(gameId) {
 
 async function getPlayersInGame(gameId) {
   const game = await repository.findGameById(gameId);
-  if (!game) throw new Error("Game not found");
+  if (!game) {
+    throw new AppError("Game not found", 404);
+  }
 
   return {
     game_id: game.id,
@@ -181,7 +209,7 @@ async function getPlayersInGame(gameId) {
 async function getCurrentPlayer(gameId) {
   const game = await repository.findGameById(gameId);
   if (!game) {
-    throw new Error("Game not found");
+    throw new AppError("Game not found", 404);
   }
 
   return {
